強化カードは日本語　済 移動速度が早い　済 壁ジャンプできちゃってる　済 溜め攻撃攻撃力=近接攻撃の1.5倍　済 チャージ時間短縮 済(3/4になった) チャージ中歩けるように　済 ジャストガードで被ダメージ0

+++ Assets/01_Scripts/SceneChanger.cs
@@ -52,7 +52,7 @@ public class SceneChanger : MonoBehaviour
             PlayerPrefs.SetInt("HP", (int)_playerMoveScripts._currentHP);
             PlayerPrefs.SetFloat("Speed", _playerMoveScripts._maxSpeed);
             PlayerPrefs.SetInt("AttackDamage", _attackScript._ATTACK_DAMAGE_MAX);
-            PlayerPrefs.SetInt("ChargeAttackDamage", _chargeAttackScript._CHARGE_ATTACK_DAMAGE_MAX);
+            PlayerPrefs.SetFloat("ChargeAttackDamage", _chargeAttackScript._CHARGE_ATTACK_DAMAGE_MAX);
             PlayerPrefs.SetFloat("BulletDamage", _playerMoveScripts._bulletDamage);
             PlayerPrefs.SetInt("MaxHP", _playerMoveScripts._maxHP);
             PlayerPrefs.Save();
